+++ server/src/save.py
@@ -18,7 +18,6 @@ SAVE_VERSION = 108
 
 import asyncio
 import json
-import logging
 import secrets
 import shutil
 import sys
@@ -32,10 +31,8 @@ from .db.db import db as ACTIVE_DB
 from .db.models.constants import Constants
 from .thumbnail import generate_thumbnail_for_asset
 from .utils import ASSETS_DIR, FILE_DIR, SAVE_PATH, OldVersionException, UnknownVersionException, get_asset_hash_subpath
-
-logger: logging.Logger = logging.getLogger("PlanarAllyServer")
-logger.setLevel(logging.INFO)
-
+# Import logger to retain configuration settings
+from .logs import logger
 
 def get_save_version(db: SqliteExtDatabase):
     return db.execute_sql("SELECT save_version FROM constants").fetchone()[0]
@@ -52,7 +49,6 @@ def create_new_db(db: SqliteExtDatabase, version: int):
         secret_token=secrets.token_bytes(32),
         api_token=secrets.token_hex(32),
     )
-
 
 def check_existence() -> bool:
     if not SAVE_PATH.exists():